+++ fuzz/fuzz_targets/decode_index_payload.rs
@@ -1,0 +1,7 @@
+#![no_main]
+
+use libfuzzer_sys::fuzz_target;
+
+fuzz_target!(|data: &[u8]| {
+    let _ = typra_core::index::decode_index_payload(data);
+});